messdeck/requirements.txt
django
django-allauth
pillow
openpyxl
pandas